Move healthchecks into a seperate template
