+++ templates/healthcheck/schema.capnp
@@ -1,0 +1,5 @@
+@0xc5ba0f64c9013a79;
+
+struct Schema {
+    node @0: Text; # reference to the node service to run the healthcheck on
+}
