+++ templates/node/node.py
@@ -90,8 +90,6 @@ class Node(TemplateBase):
         except StateCheckError:
             pass
 
-        self._healthcheck()
-
     @retry(Exception, tries=2, delay=2)
     def install(self):
         self.logger.info('Installing node %s' % self.name)
@@ -177,32 +175,6 @@ class Node(TemplateBase):
         self.state.check('status', 'running', 'ok')
         return self.node_sal.client.ping()[13:].strip()
 
-    def _healthcheck(self):
-        node_sal = self.node_sal
-        _update_healthcheck_state(self, node_sal.healthcheck.openfiledescriptors())
-        _update_healthcheck_state(self, node_sal.healthcheck.cpu_mem())
-        _update_healthcheck_state(self, node_sal.healthcheck.rotate_logs())
-        _update_healthcheck_state(self, node_sal.healthcheck.network_bond())
-        _update_healthcheck_state(self, node_sal.healthcheck.interrupts())
-        _update_healthcheck_state(self, node_sal.healthcheck.context_switch())
-        _update_healthcheck_state(self, node_sal.healthcheck.threads())
-        _update_healthcheck_state(self, node_sal.healthcheck.qemu_vm_logs())
-        _update_healthcheck_state(self, node_sal.healthcheck.network_load())
-        _update_healthcheck_state(self, node_sal.healthcheck.disk_usage())
-
-        # this is for VM. VM is not implemented yet, and we'll probably not need
-        # some cleanup like this anyhow
-        # node_sal.healthcheck.ssh_cleanup(job=job)
-
-        # TODO: this need to be configurable
-        flist_healhcheck = 'https://hub.gig.tech/gig-official-apps/healthcheck.flist'
-        with node_sal.healthcheck.with_container(flist_healhcheck) as cont:
-            _update_healthcheck_state(self, node_sal.healthcheck.node_temperature(cont))
-            _update_healthcheck_state(self, node_sal.healthcheck.powersupply(cont))
-            _update_healthcheck_state(self, node_sal.healthcheck.fan(cont))
-
-        # TODO: check network stability of  node with the rest of the nodes !
-
     def _start_all_containers(self):
         for container in self.api.services.find(template_uid=CONTAINER_TEMPLATE_UID):
             container.schedule_action('start', args={'node_name': self.name})
@@ -224,28 +196,3 @@ class Node(TemplateBase):
     def _wait_all(self, tasks, timeout=60, die=False):
         for t in tasks:
             t.wait(timeout=timeout, die=die)
-
-
-def _update(service, healcheck_result):
-    for rprtr in service.data.get('alerta', []):
-        reporter = service.api.services.get(name=rprtr)
-        reporter.schedule_action('process_healthcheck', args={'name': service.name, 'healcheck_result': healcheck_result})
-
-    category = healcheck_result['category'].lower()
-    if len(healcheck_result['messages']) == 1:
-        tag = healcheck_result['id'].lower()
-        status = healcheck_result['messages'][0]['status'].lower()
-        service.state.set(category, tag, status)
-    elif len(healcheck_result['messages']) > 1:
-        for msg in healcheck_result['messages']:
-            tag = ('%s-%s' % (healcheck_result['id'], msg['id'])).lower()
-            status = msg['status'].lower()
-            service.state.set(category, tag, status)
-
-
-def _update_healthcheck_state(service, healthcheck):
-    if isinstance(healthcheck, list):
-        for hc in healthcheck:
-            _update(service, hc)
-    else:
-        _update(service, healthcheck)
